Add prototype Python interface base class
Could also be a protocol
