+++ glue-codes/python/OpynFAST/moordyn_library.py
@@ -37,35 +37,17 @@ from ctypes import (
 import numpy as np
 import datetime
 
-class MoorDynLib(CDLL):
+from .interface_abc import OpenFASTInterfaceType
 
-    # Human readable error levels
-    error_levels = {
-        0: "None",
-        1: "Info",
-        2: "Warning",
-        3: "Severe Error",
-        4: "Fatal Error"
-    }
+class MoorDynLib(OpenFASTInterfaceType):
 
-    #   NOTE:   the error message length in Fortran is controlled by the
-    #           ErrMsgLen variable in the NWTC_Base.f90 file.  If that ever
-    #           changes, it may be necessary to update the corresponding size
-    #           here.
-    error_msg_c_len = 1025
-
-    #   NOTE:   the length of the name used for any output file written by the
-    #           HD Fortran code is 1025.
-    default_str_c_len = 1025
 
     def __init__(self, library_path):
         super().__init__(library_path)
-        self.library_path = library_path
 
         self._initialize_routines()
 
         # Create buffers for class data
-        self.abort_error_level = 4
         self.error_status_c = c_int(0)
         self.error_message_c = create_string_buffer(self.error_msg_c_len)
         self.error_message     = create_string_buffer(1025)
